proto moved to apis dir + style fixes
