+++ Makefile
@@ -19,7 +19,7 @@ VIRTUALENV_EXE=$(if $(subst 2,,$(PY_VERSION)),python3 -m venv,virtualenv)
 VIRTUALENV_DIR=$(if $(subst 2,,$(PY_VERSION)),.venv3,.venv)
 ACTIVATE="$(VIRTUALENV_DIR)/bin/activate"
 STYLEVIRTUALENV_DIR=".styleenv$(PY_VERSION)"
-STYLE_CHECK_OPTS := --exclude=ie_serving/messaging
+STYLE_CHECK_OPTS := --exclude=ie_serving/messaging/apis
 STYLE_CHECK_DIRS := tests ie_serving setup.py
 TEST_OPTS :=
 TEST_DIRS ?= tests/
